agents/worker_agent2/requirements.txt
crewai
python-dotenv
pika
langchain-openai
